# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface Certificate {
  id: number;
  title: string;
  issuer: string;
  fileUrl: string;
  /** @nullable */
  fileType?: string | null;
  category: string;
  viewCount: number;
  /** @nullable */
  shareToken?: string | null;
  uploadedByClerkId: string;
  createdAt: string;
}

export interface CertificateInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  issuer: string;
  /** @minLength 1 */
  fileUrl: string;
  fileType?: string;
  category: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface CertificateStats {
  total: number;
  totalViews: number;
  byCategory: CategoryCount[];
  mostViewed: Certificate[];
}

export interface UserProfile {
  id?: number;
  clerkId: string;
  username: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** @nullable */
  github?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  leetcode?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  createdAt?: string | null;
}

export interface UserProfileInput {
  username?: string;
  name?: string;
  bio?: string;
  profileImageUrl?: string;
  github?: string;
  linkedin?: string;
  leetcode?: string;
  resumeUrl?: string;
}

export interface Portfolio {
  user: UserProfile;
  certificates: Certificate[];
}

export type ListCertificatesParams = {
category?: string;
search?: string;
};