# lib/api-zod/src/generated/types/certificate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Certificate {
  id: number;
  title: string;
  issuer: string;
  fileUrl: string;
  /** @nullable */
  fileType?: string | null;
  category: string;
  viewCount: number;
  /** @nullable */
  shareToken?: string | null;
  uploadedByClerkId: string;
  createdAt: Date;
}