# lib/api-zod/src/generated/types/certificateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CertificateInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  issuer: string;
  /** @minLength 1 */
  fileUrl: string;
  fileType?: string;
  category: string;
}