# lib/api-zod/src/generated/types/certificateStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { Certificate } from './certificate';

export interface CertificateStats {
  total: number;
  totalViews: number;
  byCategory: CategoryCount[];
  mostViewed: Certificate[];
}