# lib/api-zod/src/generated/types/portfolio.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Certificate } from './certificate';
import type { UserProfile } from './userProfile';

export interface Portfolio {
  user: UserProfile;
  certificates: Certificate[];
}