# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id?: number;
  clerkId: string;
  username: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  profileImageUrl?: string | null;
  /** @nullable */
  github?: string | null;
  /** @nullable */
  linkedin?: string | null;
  /** @nullable */
  leetcode?: string | null;
  /** @nullable */
  resumeUrl?: string | null;
  /** @nullable */
  createdAt?: Date | null;
}