# lib/api-zod/src/generated/types/userProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Certfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfileInput {
  username?: string;
  name?: string;
  bio?: string;
  profileImageUrl?: string;
  github?: string;
  linkedin?: string;
  leetcode?: string;
  resumeUrl?: string;
}